Adjust title markup to avoid graphical glitch.
Before, the email title was italicized on mouse-over. This changed the
title width, since italic text is generally more narrow than regular.
This in turn could cause the mouse pointer to no longer hover the title,
which restored the regular text, which caused the mouse to again be over
the text, causing flickering until the mouse was moved away from the
right end of the title text.

Using bold text as general title emphasis, and signaling mouse-over by
underlining (like it is done for the action labels below), this problem
is avoided. Also, the readability is improved by using less graphical
information to signal the title -- bold face should be enough to make it
stand out clearly.