Refresh cookies on session timeout.
At 2015-03-17, my `checkgmail` session started to time out against
Google's servers every ~5 minutes, producing the login credentials
window. Just sending the credentials again refreshed the session and
made it functional for another 5 minutes.

Note that this was not always reproducible, in that it sometimes could
take 10 minutes, and sometimes a lot more. It seemed to always happen in
even 5 minute intervals, though, and in most cases it triggered on the
first such opportunity.

This change triggers a new such login request automatically when the
session times out in that fashion. It is only checked for cookie based
login, but should not be able to affect other session types. They might
need similar fixes, though.

The patch also removes a call that was already marked as "probably
unnecessary", that should safeguard against a user entering incorrect
credentials. From my testing this seems to be handled correctly without
this call, as mentioned in the earlier comments.

The code change is that the cookie login sequence has been factored out
to a separate subroutine `&cookie_login` which is called both where it
was found earlier, and automatically when cookies are activated and we
hit a 401 response after having correctly logged in once in the current
session.